+++ crates/siril-sys/src/lib.rs
@@ -8,8 +8,11 @@ pub mod message;
 pub mod siril_ext;
 
 pub use conversion::{ConversionEntry, ConversionFile};
+pub use message::SirilError;
 pub use output::{OutputLine, OutputSink, OutputStream};
 pub use siril::Builder;
+pub use siril::SIRIL_MIN_VERSION;
 pub use siril::Siril;
+pub use siril::check_siril_version;
 pub use siril::find_siril_cli;
 pub use types::*;